--- a/actors/runtime/src/main/java/cloud/orbit/actors/runtime/InvocationHandler.java
+++ b/actors/runtime/src/main/java/cloud/orbit/actors/runtime/InvocationHandler.java
@@ -34,14 +34,22 @@ import org.slf4j.LoggerFactory;
 
 import cloud.orbit.actors.Stage;
 import cloud.orbit.actors.annotation.Reentrant;
+import cloud.orbit.actors.extensions.InvocationHandlerExtension;
 import cloud.orbit.concurrent.Task;
 import cloud.orbit.util.AnnotationCache;
 
 import java.lang.reflect.Method;
+import java.util.Collections;
+import java.util.List;
+import java.util.concurrent.ConcurrentHashMap;
+import java.util.concurrent.ConcurrentMap;
+
+import static com.ea.async.Async.await;
 
 public class InvocationHandler
 {
     private static final Logger logger = LoggerFactory.getLogger(InvocationHandler.class);
+    private ConcurrentMap<Stage, List<InvocationHandlerExtension>> handlerExtensionCache = new ConcurrentHashMap<>();
 
     private final AnnotationCache<Reentrant> reentrantCache = new AnnotationCache<>(Reentrant.class);
 
@@ -100,8 +108,8 @@ public class InvocationHandler
         }
     }
 
-    @SuppressWarnings("unchecked")
-    Task<Object> invoke(Stage runtime, Invocation invocation, LocalObjects.LocalObjectEntry entry, LocalObjects.LocalObjectEntry target, ObjectInvoker invoker)
+
+    public Task<Object> invoke(final Stage runtime, final Invocation invocation, final LocalObjects.LocalObjectEntry entry, final LocalObjects.LocalObjectEntry target, final ObjectInvoker invoker)
     {
         runtime.bind();
 
@@ -133,22 +141,72 @@ public class InvocationHandler
             context.setRuntime(runtime);
         }
 
-        beforeInvoke(invocation, method);
-        final Task<Object> result = invoker.safeInvoke(target.getObject(), invocation.getMethodId(), invocation.getParams());
-        final long startTimeNanos = System.nanoTime();
-        afterInvoke(startTimeNanos, invocation, method);
+        // Perform the internal invocation
+        final Task<Object> invokeResult = internalInvoke(runtime, invocation, entry, target, method, reentrant, invoker);
+
+        // Link the result to the completion promise
         if (invocation.getCompletion() != null)
         {
-            InternalUtils.linkFutures(result, invocation.getCompletion());
+            InternalUtils.linkFutures(invokeResult, invocation.getCompletion());
         }
-        result.whenComplete((o, throwable) -> taskComplete(startTimeNanos, invocation, method));
 
+        // If reentrant we say we are done
         if (reentrant)
         {
-            // let the execution serializer proceed if actor method blocks on a task
             return Task.fromValue(null);
         }
 
-        return result;
+        return invokeResult;
+    }
+
+    @SuppressWarnings("unchecked")
+    protected Task<Object> internalInvoke(final Stage runtime, final Invocation invocation, final LocalObjects.LocalObjectEntry entry, final LocalObjects.LocalObjectEntry target, final Method method, final Boolean reentrant, final ObjectInvoker invoker)
+    {
+        final long startTimeNanos = System.nanoTime();
+        final List<InvocationHandlerExtension> invocationHandlerExtensions = getCachedHandlerExtensions(runtime);
+
+        // Before invoke actions
+        Task<?> beforeInvokeChain = Task.done();
+        for (InvocationHandlerExtension invocationHandlerExtension : invocationHandlerExtensions)
+        {
+            beforeInvokeChain = beforeInvokeChain.thenCompose(() -> invocationHandlerExtension.beforeInvoke(startTimeNanos, target.getObject(), method, invocation.getParams()));
+        }
+        await(beforeInvokeChain);
+        beforeInvoke(invocation, method);
+
+        // Real invoke
+        final Task<Object> invokeResult = invoker.safeInvoke(target.getObject(), invocation.getMethodId(), invocation.getParams());
+
+        // After invoke actions
+        afterInvoke(startTimeNanos, invocation, method);
+        Task<?> afterInvokeChain = Task.done();
+        for (InvocationHandlerExtension invocationHandlerExtension : invocationHandlerExtensions)
+        {
+            afterInvokeChain = afterInvokeChain.thenCompose(() -> invocationHandlerExtension.afterInvoke(startTimeNanos, target.getObject(), method, invocation.getParams()));
+        }
+        await(afterInvokeChain);
+
+        // After complete invoke chain actions
+        return invokeResult.thenCompose((o) ->
+        {
+            taskComplete(startTimeNanos, invocation, method);
+
+            Task<?> afterCompleteChain = Task.done();
+            for (InvocationHandlerExtension invocationHandlerExtension : invocationHandlerExtensions)
+            {
+                afterCompleteChain = afterCompleteChain.thenCompose(() -> invocationHandlerExtension.afterInvokeChain(startTimeNanos, target.getObject(), method, invocation.getParams()));
+            }
+
+            return afterCompleteChain.thenApply((f) -> o);
+        });
+    }
+
+    private List<InvocationHandlerExtension> getCachedHandlerExtensions(final Stage stage) {
+        List<InvocationHandlerExtension> invocationHandlerExtensions = handlerExtensionCache.get(stage);
+        if(invocationHandlerExtensions == null) {
+            invocationHandlerExtensions = Collections.unmodifiableList(stage.getAllExtensions(InvocationHandlerExtension.class));
+            handlerExtensionCache.putIfAbsent(stage, invocationHandlerExtensions);
+        }
+        return invocationHandlerExtensions;
     }
 }