This PR adds Invocation Handler Extensions which allow an application to implement multiple invocation handler filters/intercepts.