@JoeHegarty 2017-03-11
Slight drop which is not unexpected, but overall holds up well.

#### Before Change
Result "avgRequestTime_singleThread":
  5.950 ±(99.9%) 0.241 us/op [Average]
  (min, avg, max) = (5.169, 5.950, 8.611), stdev = 0.709
  CI (99.9%): [5.709, 6.190] (assumes normal distribution)

Run complete. Total time: 00:13:41

Benchmark                                         Mode  Cnt        Score       Error  Units
SingleNodeBenchmark.requestThroughput            thrpt  100  1693623.040 ± 36373.151  ops/s
SingleNodeBenchmark.avgRequestTime_batched        avgt  100        2.488 ±     0.046  us/op
SingleNodeBenchmark.avgRequestTime_singleThread   avgt  100        5.950 ±     0.241  us/op


#### After Change
Result "avgRequestTime_singleThread":
  5.549 ±(99.9%) 0.172 us/op [Average]
  (min, avg, max) = (5.039, 5.549, 9.087), stdev = 0.508
  CI (99.9%): [5.377, 5.722] (assumes normal distribution)

Run complete. Total time: 00:16:10

Benchmark                                         Mode  Cnt        Score       Error  Units
SingleNodeBenchmark.requestThroughput            thrpt  100  1541066.454 ± 30625.454  ops/s
SingleNodeBenchmark.avgRequestTime_batched        avgt  100        2.734 ±     0.051  us/op
SingleNodeBenchmark.avgRequestTime_singleThread   avgt  100        5.549 ±     0.172  us/op